+++ tests/UnitTest/Definition/Source/DefinitionArrayTest.php
@@ -133,9 +133,8 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
     {
         $source = new DefinitionArray();
         $definition = new ValueDefinition('value');
-        $definition->setName('foo');
 
-        $source->addDefinition($definition);
+        $source->addDefinition('foo', $definition);
         $this->assertSame($definition, $source->getDefinition('foo'));
     }
 
@@ -143,7 +142,6 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
     {
         $source = new DefinitionArray();
         $definition = new ValueDefinition('value');
-        $definition->setName('foo');
 
         $source->addDefinitions(['foo' => $definition]);
         $this->assertSame($definition, $source->getDefinition('foo'));
@@ -152,7 +150,6 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
     public function testAddDefinitionsInConstructor()
     {
         $definition = new ValueDefinition('value');
-        $definition->setName('foo');
 
         $source = new DefinitionArray(['foo' => $definition]);
         $this->assertSame($definition, $source->getDefinition('foo'));
@@ -162,9 +159,7 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
     {
         $source = new DefinitionArray();
         $definition1 = new ValueDefinition('value');
-        $definition1->setName('foo');
         $definition2 = new ValueDefinition('value');
-        $definition2->setName('foo');
 
         $source->addDefinitions(['foo' => $definition1]);
         $source->addDefinitions(['foo' => $definition2]);
@@ -183,22 +178,18 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
 
         $definition = $source->getDefinition('foo1');
         $this->assertInstanceOf(ValueDefinition::class, $definition);
-        $this->assertEquals('foo1', $definition->getName());
         $this->assertEquals('bar', $definition->getValue());
 
         $definition = $source->getDefinition('Namespaced\FooInterface');
         $this->assertInstanceOf(ObjectDefinition::class, $definition);
-        $this->assertEquals('Namespaced\FooInterface', $definition->getName());
         $this->assertEquals('Namespaced\Foo', $definition->getClassName());
 
         $definition = $source->getDefinition('Namespaced2\FooInterface');
         $this->assertInstanceOf(ObjectDefinition::class, $definition);
-        $this->assertEquals('Namespaced2\FooInterface', $definition->getName());
         $this->assertEquals('Namespaced2\Foo', $definition->getClassName());
 
         $definition = $source->getDefinition('Multiple\Foo\Bar\Matches');
         $this->assertInstanceOf(ObjectDefinition::class, $definition);
-        $this->assertEquals('Multiple\Foo\Bar\Matches', $definition->getName());
         $this->assertEquals('Multiple\Foo\Bar\Implementation', $definition->getClassName());
     }
 
@@ -213,7 +204,6 @@ class DefinitionArrayTest extends \PHPUnit_Framework_TestCase
         ]);
         $definition = $source->getDefinition('foo');
         $this->assertInstanceOf(ValueDefinition::class, $definition);
-        $this->assertEquals('foo', $definition->getName());
         $this->assertEquals('bim', $definition->getValue());
     }